requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
vector-quantize-pytorch>=1.14.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tensorboard>=2.13.0
Pillow>=9.5.0
pyyaml>=6.0
scikit-learn>=1.2.0